rag-faq-app/requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-openai
pypdf
python-dotenv
openai

